Add files via upload
removed the auto add commands to steam as it wasnt working. 
You now have colour coded commands to add your steam games.
There is a new option to show search other drives for steam games and to check dependencies and install them if they are missing , please rememebr to star the program if it helps you!